Move the @disable_ddl_transaction infos in the Migration.__migration__/0 function
Before this commit, the `Ecto.Migration.__before_compile__/1` hook defined a
`__disable_ddl_transaction__/0` function that just returned the value of the
`@disable_ddl_transaction` attribute. I removed that function in favour of
putting this information in a kw list returned by the `__migration__/0`
function. This function previously returned `true` and (as far as I understand)
was used just to check that `Ecto.Migration` had been used in a migration. As a
matter of fact, as long as this function is defined, all previous tests pass.