Don't use HashSets in Ecto.Changeset.cast/4
The previous code was using HashSets in order to merge two lists, leave only the
unique elements and then perform a set difference between this set and another
list. This has all been replaced by a call to `++/2`, a call to `Enum.uniq/1`
and a call to `--/2`. We can do this since both the first and the second lists
passed to `--/2` do not contain duplicate elements.