Make Ecto.Changeset.merge/2 public
I refactored the Ecto.Changeset.merge_changesets/2 private function into a
`merge/2` public function with the same job. This can be useful to merge
changesets that have been created in different places. This function is still
used inside `cast/4` when a changeset is passed as the second argument (instead
of a model).

I also refactored the tests in order to test `merge/2` individually.