Make Ecto.Changeset.change/2 accept changesets too
`Ecto.Changeset.change/2` now accepts a changeset as the first argument (other
than a model).

I added documentation for this change as well as updated the spec for this
function. I also slightly refactored the code for when a keyword list is passed
to `change/2` so that it converts it into a map and then calls itself.