Introduce `Ecto.Repo.after_connect/1` (refs #599)
In the current implementation `Ecto.Repo.after_connect/1` is called
synchronously with the new connection pid, after a new connection is initiated
but before the connection pid is returned to the pool.

`Ecto.Adapters.SQL` functions can't be called in the hook because the
connection is not yet returned to the pool. This means you can only used
the driver functions to make queries using the connection pid.