Minimize query rewriting on assoc query for through expressions
Imagine the following has_many through:

    defmodule User do
      has_many :applications,
        through: [:team_members, :environment, :applications]
    end

and the assoc query

    assoc(%User{id: 1}, :applications)

which first expands to

    from t in TeamMember,
      join: e in Environment,
      on: e.id == t.environment_id,
      join: a in Application,
      on: a.environment_id == e.id,
      where: t.user_id == ^1

The challenge is to rewrite the query above from the
expansion of the has_many :through to another query
that has the last join (i.e. the Application) as the
query source (from).

Prior to this commit we would rewrite as:

    from a in Application,
      join: e in Environment,
      on: e.id == t.environment_id,
      join: t in TeamMember,
      on: a.environment_id == e.id,
      where: t.user_id == ^1,
      distinct: true,
      select: a

However, this wouldn't compile because the first join is
referring to a table that does not yet exist. We have fixed
this issue by making TeamMember (the first "through") be the
first join before traversing all other "throughs" and then
by moving the Application (the last join after traversing)
become the from + where:

    from a in Application,
      join: t in TeamMember,
      on: t.user_id in ^[1],
      join: e in Environment,
      on: e.id == t.environment_id,
      where: a.environment_id == e.id,
      distinct: true,
      select: a

Closes #763