Allow anonymous functions for casting associations
This allows user's to pass a custom function for casting associations.

    cast(post, params, ~w(), [{:author, &my_custom_changeset/2}])

Also clarifies documentation on what it expects the value to be. When I
read "key value" I thought it wanted a Map or Keyword list. This commit
changes it to state that it expects a 2 item tuple and adds an example
for clarity.

This can be helpful if your changesets are defined somewhere other than
the model itself. It is also helpful even if it is defined on the model
because it makes it very clear which function is used

    cast(post, params, ~w(), [{:author, &Author.optional_changeset/2}])