Do not reorder fetch ids
Because PostgreSQL is not guaranteed to return items by ID,
if we reorder the IDs, there is no guarantee they will match
the returned items.

Notice the line of code removed was added to solve #1353 but
we can never effectively solve #1353 because the ordering will
always be dictated by the comments (the first step in the preload)
and not by the last step.

Closes #1393.