Support required has_many in traverse_errors (#1403)
A required has_many association already reports a Changeset error when
the value is empty:

    iex> changeset = %Author{}
    iex>   |> Changeset.cast(%{posts: []}, ~w())
    iex>   |> Changeset.cast_assoc(:posts, required: true)
    iex> changeset.errors
    [posts: {"can't be blank", []}]

However, calling `Changeset.traverse_errors` on this changeset would
result in a blank set of errors for `posts`:

    iex> Ecto.Changeset.traverse_errors(changeset, &(&1))
    %{posts: []}

This happened because the errors on the nested association -- of which
there were none -- overwrote the error on the parent for the association
name, i.e. it was the equivalent of the following:

    iex> Map.put(%{posts: [{"can't be blank", []}]}, :posts, [])
    %{posts: []}

The fix has been implemented by merging the nested errors only if
there is no error on the parent for the association.  I believe this is
safe because if there is an error on the association itself, we should
not care about errors on the children.