Ensure that changeset_fun always returns a Changeset in Multi
I was using insert with a changeset_fun and forgot that it must return a changeset, not a struct or map.
I got confused, because Ecto.Multi.insert without a function does accept both a changeset and struct.
When I passed in a struct that happened to contain `action: "user.create"` value, I got a really confusing error, see below.

```
** (ArgumentError) you provided a changeset with an action already set to "user.create" when trying to insert it
stacktrace:
  (ecto) lib/ecto/multi.ex:258: Ecto.Multi.put_action/2
  (ecto) lib/ecto/multi.ex:389: Ecto.Multi.apply_operation/3
  (ecto) lib/ecto/multi.ex:374: Ecto.Multi.apply_operation/4
  (elixir) lib/enum.ex:1473: Enum."-reduce/3-lists^foldl/2-0-"/3
  test/support/test_repo.exs:88: Ecto.TestAdapter.transaction/3
  (ecto) lib/ecto/repo/queryable.ex:27: Ecto.Repo.Queryable.transaction/4
```

This PR adds some guards and a friendly error.