Fixes embed changeset corruption when parent insert fails.
This is fixes the bug reported in #1584. The issue was caused by
the fact that embedded changes are already replaced in the changeset
with their target struct before the parent records in the changeset
are processed. So if a parent insert causes an error, we end up
returning a changeset with the replaced struct, not the original
changeset for the embedded object. I've fixed this by checking if
the changeset is still valid after attempting to process the
the parents, and only after that check are the embed changesets
converted to structs.