Update Travis CI config to use Postgres 9.5
The UPSERT feature that is used by several integration tests is
only available in Postgres 9.5.

Unfortunately, the only environment that supports 9.5 is the trusty
distribution using sudo. Enabling this for now to ensure that the
tests pass with 9.5 available, can revert back if there are better
alternatives.

https://github.com/travis-ci/travis-ci/issues/4264
https://docs.travis-ci.com/user/database-setup/#Using-a-different-PostgreSQL-Version

Also, MySQL is no longer available as a service. The recommended
way to install is via `apt` packages.

https://docs.travis-ci.com/user/database-setup/#MySQL-5.6
https://github.com/travis-ci/docs-travis-ci-com/issues/396#issuecomment-156507245