Allow cast_assoc to work inside embeds
The issue is that embeds don't have the __meta__ field, so we want to
raise only for regular schemas.
This, unfortunately, creates another problem. We don't know if the embed
was just created or loaded from database and needs preloading the
association data. This is very risky to use as it may cause data loss
if used improperly.