Switch id fields to default to BIGINT
This change explicitly aligns PG and MySQL implementations to utilise
bigint fields for primary and foreign keys. Previously this relied on
falling through to the definition of serial supplied by the database
engine which differed:

- MySQL resolved serial to `BIGINT UNSIGNED NOT NULL AUTO_INCREMENT
  UNIQUE`
- PostgreSQL resolved serial to an auto incrementing integer with a
  maximum value of 2147483647

n.b. this commit no longer relies on the MySQL definition of serial
as it marks the field UNIQUE which differs to the behaviour within PG.
With this implementation both adaptors should exhibt the same behaviour.

Fixes elixir-ecto/ecto#1879