Define internal type bigserial
Instead of overloading the semantics of the existing serial type
we introduce a bigserial type and utilise that by default within new
migrations.

The definition of serial is made explicit here but it keeps the
semantics MySQL already has associated with it, namely it was already
using the BIGINT type.