Use the "postgres" database for create/drop database commands
Formerly, `mix ecto.create` or `mix ecto.drop` used the "template1" database.
However, other processes are not able to create new databases as long as there
is someone connected to the "template1" database. To fix that, we are now
connecting to the "postgres" database, which is also what PostgreSQL tools like
`createdb` do.