Add apply_action/3 to changeset
Useful for composition in Ecto.Multi's to apply
changes only when valid and return an error
tuple in the event of invalid changes.
Additionally, an action must be provied to
represent the action being peformed from the
changes. This is useful when working with embedded
schemas where operations on the changeset are
committed manually without relying on a Repo to
reflect failed operations on the changeset by
updating the action field.