Fixup docs for schemaless changesets (#1943)
"However, changesets also run without a changeset" is a bit self-referential, so I've tried to distinguish between data in a struct and data in a map (which has no implied schema). Also, the "as a tuple instead of a struct" bit seems superfluous (it's like saying "pass a tuple as a tuple").