Add tests
Please note that raw commands sequenced by execute/1 and execute/2 are treated
differently from other commands.

Commands to create a table, constraint or index work in a two-step way,
by creating a struct holding a representation of the object to be created
and then by passing it to the runner by explicitly calling drop or create.

Raw commands go straight to the Runner, so it is not possible to test
them in the same manner; the difference is visible in the test case,
because it must test for an :ok from the runner.