Allow create table to optional add unlogged
This is only applicable to Postgres adaptor.
UNLOGGED optiona prevents Postgres using the WAL :

> Data written to unlogged tables is not written to the write-ahead log
> which makes them considerably faster than ordinary tables.
- https://www.postgresql.org/docs/9.1/static/sql-createtable.html

Problem with this commit is that by adding to Ecto.Migration.Table
means it needs to work on all adapters.
Cannot use options either as that raises if keyword list is used,
as its used for generic text to add to column SQL commands.

Ideally this would work :

    create table(:unlogged_true, options: [unlogged: true])