Extract load_structure function from load mix task
Intent is to wrap the structure loading code in a conditional so it only executes if he user provides confirmation. This way we can wrap the conditional around a single function call. I think that will be clearer than wrapping a conditional around the code in place. See https://github.com/elixir-ecto/ecto/issues/2144