Allow validation of codepoint count in validate_length
A common use case of validate_length is probably to check
that the given text will fit into a text field with a limited
size.

Most databases ignore graphemes in the size limit for char fields,
and only care about the number of code points.

This commit introduces the `:count` option for `validate_length`,
which will count codepoints instead of graphemes.