Correct the description of how many_to_many works (#2188)
put_assoc raises a runtime error when changing many to many relationships.  In the example provided in the docs, removing the "clickbait" tag will raise an error.

To make the removal work as originally presented in the documentation, one must first change the on_replace behaviour from its default of :raise to :delete.