Improve transaction handling
Transactions now mark SQL transactions as failed if attempt to
begin, rollback or commit a transaction when SQL status does not
allow it.

Nested transaction now mark SQL transactions as failed if a nested
transaction starts or ends outside a transaction.

Nested transactions now mark pg transactions as failed when a SQL
error occurs that put transaction in error state without an
uncaught exception or explicit rollback to notify Ecto.