Add `migration_default_prefix` to repo config
When you are working with a database that *always* uses an alternate
schema prefix to `public`, allow setting a default for all migrations
to use. From the model standpoint its easy to set @schema_prefix in a
module then use that module in your models, from the migration
perspective you have to remember to run migrations with `--prefix` or
set the prefix on each DDL statement which is prone to forgetfulness :D

The default is overridable at the DDL level by specifying
`prefix: "foo"` on a create table, index, or drop statement.

It is also overridable by passing `--prefix NAME` to the runner.