Add fragment support for conflict targets
This is to clean up a remaining bug from https://github.com/elixir-ecto/ecto/issues/2081.

Basically, Postgres unique indexes support expressions and partial indexing, both of which make it
impossible to define an explicit constraint for them.  Ecto conflict targets on upserts
automagically columnize each clause, and the constraint alternative also is no help, because no constraint
can be made for such a unique index, so this is a simple stopgap to handle these corner cases.