Properly match keyword lists on query planner.
Fixes an error that happens when you try to query a list of custom ecto type which is defined as a tuple, but it's not a kw list. Eg:

```
locations = [{"europe", "france"}, {"south_america", "brazil"}]
from(c in City, where: [location: locations])
```

In that case `location` would be a custom ecto type that cast/dump to string, jsonb, etc.