No need for validation query unless relevant fields were changed
unsafe_validate_unique/3 performs a query to ensure that the specified
fields represent a unique combination of values in the repo. For
example, perhaps the combination of city and state must be unique.
But there's no need to perform this query if the changeset is not making
any change to city or state.

Fixes #2901