Expose the type that was validated in errors generated by `Changeset.validate_length/3` (#2913)
Errors in an `Ecto.Changeset` have a key, a message and additional
contextual information provided in the form of a keyword list.

`Ecto.Changeset` comes bundled with a number of validations,
for example `validate_length/3`.

The built-in validations tend to include additional contextual
information - for example, `validate_length/3` includes `count`
and `kind`. So, for example, if the validation was configured to
check that a field was "no more than 3" and it fails, then you will
have `[count: 3, validation: :length, kind: :max]`.

The `validate_length` validation is interesting, because it can
operate on strings, binaries and lists. However, it doesn't expose
in the additional context what type of data was validated.

This context can be valuable, for example if you want your errors
to be fully machine-readable.

This exposes `[type: :string]`, `[type: :list]` or `[type: :binary]`,
as appropriate, on the error.