Add tests and better documentation for behavior around `prefix` option
There was some confusing documentation and very little test coverage
around the behavior of the `prefix` option in the `Ecto.Repo` behaviour.
I've added more tests to clarify the current expected behavior, and
documented that behavior more thoroughly so folks aren't surprised by
some of the edge cases that arrise when using that option.