Add `not_equal_to` option for `validate_number` (#2923)
This checks for inequality with given parameter. It is useful in
situations when we want to disallow one particular value. For example
imagine that we want to define table to store value income and outcome
in some account. It make sense to be positive (income) or negative
(outcome), but it makes no sense to store operation that has value equal
to `0`, with this change we can simply:

```elixir
validate_number(changeset, :diff, not_equal_to: 0)
```