Expose the enum that was validated against in errors generated by `Changeset.validate_exclusion/4`
Errors in an `Ecto.Changeset` have a key, a message and additional
contextual information provided in the form of a keyword list.

`Ecto.Changeset` comes bundled with a number of validations, for
example `validate_inclusion/4`.

The built-in validations tend to include additional contextual
information - for example, `validate_length/3` includes `count`
and `kind`. So, for example, if the validation was configured to
check that a field was "no more than 3" and it fails, then you
will have `[count: 3, validation: :length, kind: :max]`.

There are a number of validations that validate a field against
an object that implements the `Enumerable` protocol - for example
`validate_inclusion/4` which checks that the field's value is
included in a supplied enumerable.

However, these validations don't expose in the additional context
the enumerable that the value was validated against.

This context can be valuable, for example if you want your errors
to be fully machine-readable.

This exposes `[enum: enum_that_the_field_was_validated_against]`
on errors generated by `validate_exclusion/4`.