Fix: prefix for many to many associations
Based on this bug report: https://github.com/elixir-ecto/ecto/issues/2985

`join_through` for many to many associations can be defined using a
schema or a table name. The problem was that the `rewrite_join_prefix`
method for the second case has not been defined. Because of that the
prefix assignment was skipped.