Don't wrap in a transaction if we're already in a transaction
The implicit rollback in wrap_in_transaction works well if the transaction
is created by wrap_in_transaction itself, but if the function
is called from a pre-existing transaction, it takes over the rollback
decision from the parent context.

This change enables savepoint recovery to work following a failed
insert that has embeds/relations, which is implicitly wrapped into a
transaction.

Transactional integrity should not be impacted, as in the case of an
error, any further calls in the transaction prior to rollback will
be rejected even with this change.

Further discussion in elixir-ecto/ecto_sql#136.