Fixing issues in code examples
On `insert_and_get_all_tags()`, inside the case, for consistency, it should expect `names`. If you put `tags`, it will give an error (is not expecting it). Or you can change all to tags, could make more sense semantically speaking.

For some reason, which I don't know why yet, `timestamps()` expects that you truncate the second of the `NativeDateTime` given format. If not, it will give you an error.