Fix phrasing in custom type example
The phrasing of this example is misleading. It implies that the database
will ensure we get back a map, which is wrong. Ecto or its adapters
ensure we get back a map when using Ecto's :map type. However, in
practice, adapters use column types that are not that strict to store
:map, like jsonb in PostgreSQL, which can contain any valid JSON value
(e.g. a boolean, a string or an array).

The is_map guard in this example ensures we get back a map, not the
database.