Preloader executes only queries in parallel
Prior to this patch, the preloader would execute the whole
association subtree in parallel. For example, if posts has
many comments that has many votes and we were preloading
multiple entries at those levels, we would start tasks to
preload comments and friends which then would start tasks
to preloads all votes and friends. This lead to a bunch of
data copying between parent-child tasks.

The current implementation makes it so only queries are
executed in parallel. So for the case above, we will preload
all comments and friends in tasks which then return to the
main task, which then preloads all votes and friends.

Closes #3295.