This function puts a change `new_key` only if a value exists in the changeset for an existing `key`.
I have noticed that `Ecto.Changeset.get_change/3` is used very frequently in codebases to test for the presence of an existing change in a changeset, and if the change is present, to put some _other_ change, often conditional on the value of the existing change.

In practice this looks like:
```elixir
case get_change(changeset, :key) do
    nil -> changeset
    value -> put_change(changeset, :new_key, new_value)
end
```

A convenient function to put a new change if an existing one is present would be a small but highly useful addition.