Align behaviour cast for { :array, :string } with :string
In the refactor of type.ex for 3.5 an breaking change and
inconsistency was introduced:

Prior to the Type refactor:
```
3.4> Ecto.Type.cast( {:array, :string} , ["hello" , nil ])
{:ok, ["hello", nil]}
```

After the Type refactor:
```
3.5> Ecto.Type.cast( {:array, :string} , ["hello" , nil ])
:error
```

It's also an inconsistency with a cast invocation like this:
```
3.5> Ecto.Type.cast( :string, nil )
{:ok, nil}
```

After amending type.ex with this head:
```
 defp array([nil | t], fun, acc ),
    do: array( t, fun, [nil | acc ] )
```

which would resolve the `{ :array, :string }` problem,
a few test cases fail, an :integer array [3,nil] which
is a 3.5 `:error` now returns [3,nil]. The parameterized
test exhibits similar behaviour. The `:integer` behaviour
mirror the :string change.

I do believe the above array() head is the correct
and consistent behaviour given `cast( :integer, nil ) -> { :ok, nil }`.