fix: parent_as when used at right side of a combining expression
This patch fixes a bug in which is impossible to compile a query using
`parent_as` at the right side of a union expression (or any other type
of combination). As an example, consider the following query:

```
  subquery = from(c in Comments, where: parent_as(:posts).id == c.post_id)
  from(p in Posts, where: exists(union_all(subquery, ^subquery)))
```

This query should fail to compile with the following error:

```
** (Ecto.SubQueryError) the following exception happened when compiling a subquery.
    ** (Ecto.QueryError) `parent_as(:post)` can only be used in subqueries in query:
```

The problem, as it seems, is that the `plan_subquery` does not include
the required alias on the combination queries -- the right side query
of the union operator in this case. Thus, when the times come to
resolve the alias, it fails with the above message.

To fix this issue, the function `plan_subquery` has been modified to
include the `@parent_as` alias for every elemen in the `combination`
field. This is done recursevily because the query may have two or more
combining queries.

There is also a required change on ecto_adapters repository to make
this solution works, which I'll be submiting patches for as well.