Raise on duplicate values in Ecto.Enum
Duplicate values make so that only the last of the duplicates is taken
into account, due to the fact that the keyword list is cast into a map
for the `on_load` and `on_dump` values. With this commit we can avoid
this scenario in compile time.

Pointed out by @wojtekmach on #3676