Update aborted transaction docs
It seems I have overlooked the savepoints feature in https://github.com/elixir-ecto/ecto/pull/3946 and it's actually an SQL standard. It seems it's currently supported by all our default adapters, so we shouldn't make the case that it's specific to Postgres.

BTW, Postgrex has this option documented here https://hexdocs.pm/postgrex/Postgrex.html#transaction/3, but I had to look up the source code to find that MyXQL and TDS also support the savepoints feature. Should we document this as part of Ecto `Repo.transaction/2` as well?