Ensure that join_where conditions preload correctly
Every other place that `Ecto.Association.combine_joins_query` is called,
the `Ecto.Query` it is called with contains `joins` against a fixed
"base" queryable (usually `owner`). In those cases, a fixed value of
`binding` is correct.

However, in the specific case of:
* a many-to-many with a `join_where` clause
* being preloaded with a query that contains one or many `joins`

the actual `binding` value desired is "the last one added" in `assoc_query`,
so that the `join_where` query correctly references the join table.

Without this change, the `TestRepo.preload` call in the new test fails with
the message:

```
** (Ecto.QueryError) lib/ecto/association.ex:1270: field `public` in `join` does not exist in schema Ecto.Repo.ManyToManyTest.SubAssoc in query:

from m0 in Ecto.Repo.ManyToManyTest.MyAssoc,
  join: s1 in Ecto.Repo.ManyToManyTest.SubAssoc,
  on: s1.my_assoc_id == m0.id and s1.y == ^"foo",
  join: s2 in "schemas_assocs",
  on: m0.id == s2.my_assoc_id and s1.public == ^true,
  where: s2.my_schema_id in ^[1],
  order_by: [asc: type(s2.my_schema_id, :id)],
  select: {type(s2.my_schema_id, :id), m0}
```