Simplify the handling of empty values
Prior to this patch, the logic for empty values
was scatted both on cast and on validate_required/2
(which would check for empty trimmed strings).

This PR adds support for functions in empty_values,
which allows us to move the prunning from
validate_required to cast.

On the update notes, we should advise those passing
`empty_values` to build on top of
`Ecto.Changeset.empty_values/1`.