Fix docs on :prefix option for Repo Queryable callbacks
We noticed that the docs for `c:Ecto.Repo.update_all/3` and
`c:Ecto.Repo.delete_all/3` read

> This overrides the prefix set in the query and any
> `@schema_prefix` set in the schema.

This is not true, `update_all` and `delete_all` follow the query prefix
precendence as the other functions in `Queryable`.

We also added a fourth precendence rule to point out that an already
set query prefix (e.g. applied with `put_query_prefix/2`) takes
precendence over the `:prefix` option of the `Repo` function.