docs: add clarification and examples to `Ecto.UUID.cast/1`
This highlights the fact that it can take a 16-byte binary (naturally, even in
the form of a string) which may be unexpected to users. It also adds some
recommendations for `Ecto.UUID.load/1` as a possible alternative for some uses.

See the following issues for more context and details:

- https://github.com/elixir-ecto/ecto/issues/3125
- https://github.com/elixir-ecto/ecto/issues/4174
- https://github.com/elixir-ecto/ecto/issues/4224